+++ src/zoid/card-fields/component.jsx
@@ -25,6 +25,7 @@ import {
   getUserIDToken,
   getSDKToken,
   getClientMetadataID,
+  isPayPalDomain,
 } from "@paypal/sdk-client/src";
 import { getRefinedFundingEligibility } from "@paypal/funding-components/src";
 import {
@@ -34,7 +35,7 @@ import {
   type FundingEligibilityType,
 } from "@paypal/sdk-constants/src";
 
-import { getSessionID } from "../../lib";
+import { getSessionID, ValidationError } from "../../lib";
 
 import { CardPrerender } from "./prerender";
 
@@ -90,6 +91,7 @@ type CardFieldsProps = {|
     onInputSubmitRequest?: () => ZalgoPromise<Object> | Object,
   |},
   createOrder: () => ZalgoPromise<string> | string,
+  createSubscription?: () => ZalgoPromise<string> | string,
   createVaultSetupToken: () => ZalgoPromise<string>,
   onApprove: (
     {| returnUrl?: string, vaultSetupToken?: string |},
@@ -109,6 +111,7 @@ type CardFieldsProps = {|
   hcfSessionID: string,
   partnerAttributionID: string,
   merchantID: $ReadOnlyArray<string>,
+  sdkToken?: string,
   installments?: {|
     onInstallmentsRequested: () =>
       | InstallmentsConfiguration
@@ -244,6 +247,24 @@ export const getCardFieldsComponent: () => CardFieldsComponent = memoize(
             required: false,
             value: ({ props }) => props.parent.props.createOrder,
           },
+
+          ...(isPayPalDomain() && {
+            createSubscription: {
+              type: "function",
+              required: false,
+              value: ({ props }) => {
+                if (
+                  props.parent.props.createSubscription &&
+                  !props.parent.props.sdkToken
+                ) {
+                  throw new ValidationError(
+                    `SDK Token must be passed in for createSubscription`
+                  );
+                }
+                return props.parent.props.createSubscription;
+              },
+            },
+          }),
 
           createVaultSetupToken: {
             type: "function",
@@ -572,6 +593,21 @@ export const getCardFieldsComponent: () => CardFieldsComponent = memoize(
           type: "function",
           required: false,
         },
+
+        ...(isPayPalDomain() && {
+          createSubscription: {
+            type: "function",
+            required: false,
+            value: ({ props }) => {
+              if (props.createSubscription && !props.sdkToken) {
+                throw new ValidationError(
+                  `SDK Token must be passed in for createSubscription`
+                );
+              }
+              return props.createSubscription;
+            },
+          },
+        }),
 
         createVaultSetupToken: {
           type: "function",